Cleaning up the warnings given by MSVC

+++ src/NovelRT/Animation/SpriteAnimator.cpp
@@ -46,7 +46,7 @@ namespace NovelRT::Animation {
         }
 
 
-        _accumulatedDelta += delta;
+        _accumulatedDelta += static_cast<float>(delta);
         break;
       }
     }
